+++ react-hot-button-component/src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import CustomButton from './hot-button';
+import CustomButton from './hotButton';
 
 ReactDOM.render(
   <CustomButton />,
